feat(deep-review): add adaptive review runtime
