+++ src/crates/core/src/agentic/deep_review/mod.rs
@@ -1,0 +1,16 @@
+//! Deep Review runtime policy modules and tool adapters.
+
+pub mod budget;
+pub mod concurrency_policy;
+pub mod constants;
+pub mod diagnostics;
+pub mod execution_policy;
+pub mod incremental_cache;
+pub mod manifest;
+pub mod queue;
+pub mod report;
+pub mod shared_context;
+pub mod task_adapter;
+pub mod team_definition;
+pub mod tool_context;
+pub mod tool_measurement;
