+++ src/crates/core/src/agentic/tools/implementations/task_tool.rs
@@ -1,21 +1,21 @@
 use crate::agentic::agents::{get_agent_registry, AgentInfo};
 use crate::agentic::coordination::get_global_coordinator;
+use crate::agentic::deep_review::task_adapter::{
+    self as deep_review_task_adapter, DeepReviewProviderQueueWaitOutcome,
+    DeepReviewQueueWaitOutcome, DeepReviewQueueWaitSkipReason,
+};
 use crate::agentic::deep_review_policy::{
-    classify_deep_review_capacity_error, clear_deep_review_queue_control_for_tool,
     deep_review_active_reviewer_count, deep_review_effective_concurrency_snapshot,
     deep_review_effective_parallel_instances, deep_review_has_judge_been_launched,
-    deep_review_max_retries_per_role, deep_review_queue_control_snapshot,
-    load_default_deep_review_policy, record_deep_review_capacity_skip,
-    record_deep_review_effective_concurrency_capacity_error,
-    record_deep_review_effective_concurrency_success, record_deep_review_runtime_queue_wait,
+    deep_review_turn_elapsed_seconds, load_default_deep_review_policy,
+    record_deep_review_effective_concurrency_success, record_deep_review_runtime_auto_retry,
+    record_deep_review_runtime_auto_retry_suppressed, record_deep_review_runtime_manual_retry,
     record_deep_review_task_budget, try_begin_deep_review_active_reviewer,
     DeepReviewActiveReviewerGuard, DeepReviewCapacityQueueReason, DeepReviewConcurrencyPolicy,
     DeepReviewExecutionPolicy, DeepReviewIncrementalCache, DeepReviewPolicyViolation,
     DeepReviewRunManifestGate, DeepReviewSubagentRole, DEEP_REVIEW_AGENT_TYPE,
 };
-use crate::agentic::events::{
-    DeepReviewQueueReason, DeepReviewQueueState, DeepReviewQueueStatus, ErrorCategory,
-};
+use crate::agentic::events::DeepReviewQueueStatus;
 use crate::agentic::tools::framework::{
     Tool, ToolRenderOptions, ToolResult, ToolUseContext, ValidationResult,
 };
@@ -25,35 +25,13 @@ use crate::util::errors::{BitFunError, BitFunResult};
 use async_trait::async_trait;
 use log::warn;
 use serde_json::{json, Value};
-use std::collections::{HashMap, HashSet};
+use std::collections::HashMap;
 use std::path::Path;
-use tokio::time::{sleep, Duration, Instant};
 
 pub struct TaskTool;
 
 const LARGE_TASK_PROMPT_SOFT_LINE_LIMIT: usize = 180;
 const LARGE_TASK_PROMPT_SOFT_BYTE_LIMIT: usize = 16 * 1024;
-#[cfg(test)]
-const DEEP_REVIEW_QUEUE_POLL_INTERVAL: Duration = Duration::from_millis(10);
-#[cfg(not(test))]
-const DEEP_REVIEW_QUEUE_POLL_INTERVAL: Duration = Duration::from_secs(1);
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum DeepReviewQueueWaitSkipReason {
-    QueueExpired,
-    UserCancelled,
-    OptionalSkipped,
-}
-
-enum DeepReviewQueueWaitOutcome {
-    Ready {
-        guard: DeepReviewActiveReviewerGuard<'static>,
-    },
-    Skipped {
-        queue_elapsed_ms: u64,
-        skip_reason: DeepReviewQueueWaitSkipReason,
-    },
-}
 
 impl Default for TaskTool {
     fn default() -> Self {
@@ -66,168 +44,40 @@ impl TaskTool {
         Self
     }
 
-    fn string_for_any_key<'a>(value: &'a Value, keys: &[&str]) -> Option<&'a str> {
-        keys.iter().find_map(|key| {
-            value
-                .get(*key)
-                .and_then(Value::as_str)
-                .map(str::trim)
-                .filter(|value| !value.is_empty())
-        })
-    }
-
-    fn value_for_any_key<'a>(value: &'a Value, keys: &[&str]) -> Option<&'a Value> {
-        keys.iter().find_map(|key| value.get(*key))
-    }
-
-    fn u64_for_any_key(value: &Value, keys: &[&str]) -> Option<u64> {
-        keys.iter()
-            .find_map(|key| value.get(*key).and_then(Value::as_u64))
-    }
-
-    fn string_array_for_any_key(
-        value: &Value,
-        keys: &[&str],
-    ) -> Result<Vec<String>, DeepReviewPolicyViolation> {
-        let Some(array) = Self::value_for_any_key(value, keys).and_then(Value::as_array) else {
-            return Err(DeepReviewPolicyViolation::new(
-                "deep_review_retry_missing_coverage",
-                format!("Retry coverage requires array field '{}'", keys[0]),
-            ));
-        };
-
-        let mut result = Vec::with_capacity(array.len());
-        for item in array {
-            let Some(path) = item.as_str().map(str::trim).filter(|path| !path.is_empty()) else {
-                return Err(DeepReviewPolicyViolation::new(
-                    "deep_review_retry_invalid_coverage",
-                    format!(
-                        "Retry coverage field '{}' must contain non-empty strings",
-                        keys[0]
-                    ),
-                ));
-            };
-            result.push(path.to_string());
-        }
-
-        Ok(result)
-    }
-
-    fn work_packets_from_manifest(run_manifest: Option<&Value>) -> Option<&Vec<Value>> {
-        run_manifest?
-            .get("workPackets")
-            .or_else(|| run_manifest?.get("work_packets"))?
-            .as_array()
-    }
-
-    fn packet_id_from_description(description: Option<&str>) -> Option<String> {
-        let description = description?;
-        let start = description.find("[packet ")? + "[packet ".len();
-        let packet_id = description[start..].split(']').next()?.trim();
-        (!packet_id.is_empty()).then(|| packet_id.to_string())
-    }
-
-    fn packet_belongs_to_subagent(packet: &Value, subagent_type: &str) -> bool {
-        Self::string_for_any_key(
-            packet,
-            &["subagentId", "subagent_id", "subagentType", "subagent_type"],
-        )
-        .is_some_and(|value| value == subagent_type)
-    }
-
-    fn packet_id_for_manifest_packet(packet: &Value) -> Option<&str> {
-        Self::string_for_any_key(packet, &["packetId", "packet_id"])
-    }
-
     fn deep_review_packet_id_for_cache(
         subagent_type: &str,
         description: Option<&str>,
         run_manifest: Option<&Value>,
     ) -> Option<String> {
-        let packets = Self::work_packets_from_manifest(run_manifest)?;
-
-        if let Some(description_packet_id) = Self::packet_id_from_description(description) {
-            return packets
-                .iter()
-                .any(|packet| {
-                    Self::packet_id_for_manifest_packet(packet)
-                        .is_some_and(|packet_id| packet_id == description_packet_id)
-                        && Self::packet_belongs_to_subagent(packet, subagent_type)
-                })
-                .then_some(description_packet_id);
-        }
-
-        let mut matches = packets.iter().filter_map(|packet| {
-            if Self::packet_belongs_to_subagent(packet, subagent_type) {
-                Self::packet_id_for_manifest_packet(packet).map(str::to_string)
-            } else {
-                None
-            }
-        });
-        let packet_id = matches.next()?;
-        if matches.next().is_some() {
-            None
-        } else {
-            Some(packet_id)
-        }
+        deep_review_task_adapter::deep_review_packet_id_for_cache(
+            subagent_type,
+            description,
+            run_manifest,
+        )
     }
 
     fn attach_deep_review_cache(run_manifest: &mut Value, cache_value: Option<Value>) {
-        if run_manifest.get("deepReviewCache").is_some() {
-            return;
-        }
-        let Some(cache_value) = cache_value else {
-            return;
-        };
-        if let Some(object) = run_manifest.as_object_mut() {
-            object.insert("deepReviewCache".to_string(), cache_value);
-        }
+        deep_review_task_adapter::attach_deep_review_cache(run_manifest, cache_value);
     }
 
     fn deep_review_retry_guidance_max_retries(
         effective_policy: Option<&DeepReviewExecutionPolicy>,
         dialog_turn_id: &str,
     ) -> usize {
-        effective_policy
-            .map(|policy| policy.max_retries_per_role)
-            .unwrap_or_else(|| deep_review_max_retries_per_role(dialog_turn_id))
-    }
-
-    fn manifest_packet_by_id<'a>(
-        run_manifest: Option<&'a Value>,
-        packet_id: &str,
-        subagent_type: &str,
-    ) -> Option<&'a Value> {
-        Self::work_packets_from_manifest(run_manifest)?
-            .iter()
-            .find(|packet| {
-                Self::packet_id_for_manifest_packet(packet).is_some_and(|id| id == packet_id)
-                    && Self::packet_belongs_to_subagent(packet, subagent_type)
-            })
-    }
-
-    fn assigned_scope_files_for_packet(
-        packet: &Value,
-    ) -> Result<Vec<String>, DeepReviewPolicyViolation> {
-        let Some(scope) = Self::value_for_any_key(packet, &["assignedScope", "assigned_scope"])
-        else {
-            return Err(DeepReviewPolicyViolation::new(
-                "deep_review_retry_missing_packet_scope",
-                "Retry source packet is missing assigned_scope",
-            ));
-        };
-        Self::string_array_for_any_key(scope, &["files"])
-    }
-
-    fn is_retryable_capacity_reason(reason: &str) -> bool {
-        matches!(
-            reason,
-            "local_concurrency_cap"
-                | "provider_rate_limit"
-                | "provider_concurrency_limit"
-                | "retry_after"
-                | "temporary_overload"
+        deep_review_task_adapter::deep_review_retry_guidance_max_retries(
+            effective_policy,
+            dialog_turn_id,
         )
+    }
+
+    fn should_emit_deep_review_retry_guidance(
+        is_partial_timeout: bool,
+        is_retry: bool,
+        deep_review_subagent_role: Option<DeepReviewSubagentRole>,
+    ) -> bool {
+        is_partial_timeout
+            && !is_retry
+            && matches!(deep_review_subagent_role, Some(DeepReviewSubagentRole::Reviewer))
     }
 
     fn ensure_deep_review_retry_coverage(
@@ -235,223 +85,74 @@ impl TaskTool {
         subagent_type: &str,
         run_manifest: Option<&Value>,
     ) -> Result<Vec<String>, DeepReviewPolicyViolation> {
-        let Some(coverage) = Self::value_for_any_key(input, &["retry_coverage", "retryCoverage"])
-        else {
-            return Err(DeepReviewPolicyViolation::new(
-                "deep_review_retry_missing_coverage",
-                "DeepReview retry requires structured retry_coverage metadata",
-            ));
-        };
-        let packet_id = Self::string_for_any_key(coverage, &["source_packet_id", "sourcePacketId"])
-            .ok_or_else(|| {
-                DeepReviewPolicyViolation::new(
-                    "deep_review_retry_missing_packet_id",
-                    "DeepReview retry coverage requires source_packet_id",
-                )
-            })?;
-        let source_status = Self::string_for_any_key(coverage, &["source_status", "sourceStatus"])
-            .ok_or_else(|| {
-                DeepReviewPolicyViolation::new(
-                    "deep_review_retry_missing_status",
-                    "DeepReview retry coverage requires source_status",
-                )
-            })?;
-        match source_status {
-            "partial_timeout" => {}
-            "capacity_skipped" => {
-                let capacity_reason =
-                    Self::string_for_any_key(coverage, &["capacity_reason", "capacityReason"])
-                        .unwrap_or_default();
-                if !Self::is_retryable_capacity_reason(capacity_reason) {
-                    return Err(DeepReviewPolicyViolation::new(
-                        "deep_review_retry_non_retryable_status",
-                        format!(
-                            "DeepReview retry cannot redispatch non-transient capacity reason '{}'",
-                            capacity_reason
-                        ),
-                    ));
-                }
-            }
-            other => {
-                return Err(DeepReviewPolicyViolation::new(
-                    "deep_review_retry_non_retryable_status",
-                    format!(
-                        "DeepReview retry only supports partial_timeout or transient capacity failures, not '{}'",
-                        other
-                    ),
-                ));
-            }
-        }
-
-        let packet = Self::manifest_packet_by_id(run_manifest, packet_id, subagent_type)
-            .ok_or_else(|| {
-                DeepReviewPolicyViolation::new(
-                    "deep_review_retry_unknown_packet",
-                    format!(
-                        "DeepReview retry source packet '{}' does not match reviewer '{}'",
-                        packet_id, subagent_type
-                    ),
-                )
-            })?;
-        let original_files = Self::assigned_scope_files_for_packet(packet)?;
-        Self::ensure_deep_review_retry_timeout(input, packet)?;
-        let retry_scope_files =
-            Self::string_array_for_any_key(coverage, &["retry_scope_files", "retryScopeFiles"])?;
-        let covered_files =
-            Self::string_array_for_any_key(coverage, &["covered_files", "coveredFiles"])?;
-        if retry_scope_files.is_empty() {
-            return Err(DeepReviewPolicyViolation::new(
-                "deep_review_retry_empty_scope",
-                "DeepReview retry requires at least one retry_scope_files entry",
-            ));
-        }
-
-        let original_file_set: HashSet<&str> = original_files.iter().map(String::as_str).collect();
-        let mut retry_file_set = HashSet::new();
-        for file in &retry_scope_files {
-            if !retry_file_set.insert(file.as_str()) {
-                return Err(DeepReviewPolicyViolation::new(
-                    "deep_review_retry_duplicate_scope_file",
-                    format!("DeepReview retry scope repeats file '{}'", file),
-                ));
-            }
-            if !original_file_set.contains(file.as_str()) {
-                return Err(DeepReviewPolicyViolation::new(
-                    "deep_review_retry_scope_outside_packet",
-                    format!(
-                        "DeepReview retry file '{}' is outside source packet '{}'",
-                        file, packet_id
-                    ),
-                ));
-            }
-        }
-        if retry_scope_files.len() >= original_files.len() {
-            return Err(DeepReviewPolicyViolation::new(
-                "deep_review_retry_scope_not_reduced",
-                "DeepReview retry_scope_files must be smaller than the source packet scope",
-            ));
-        }
-
-        for file in &covered_files {
-            if !original_file_set.contains(file.as_str()) {
-                return Err(DeepReviewPolicyViolation::new(
-                    "deep_review_retry_coverage_outside_packet",
-                    format!(
-                        "DeepReview retry covered file '{}' is outside source packet '{}'",
-                        file, packet_id
-                    ),
-                ));
-            }
-            if retry_file_set.contains(file.as_str()) {
-                return Err(DeepReviewPolicyViolation::new(
-                    "deep_review_retry_coverage_overlaps_scope",
-                    format!(
-                        "DeepReview retry covered file '{}' cannot also be in retry_scope_files",
-                        file
-                    ),
-                ));
-            }
-        }
-
-        Ok(retry_scope_files)
+        deep_review_task_adapter::ensure_deep_review_retry_coverage(
+            input,
+            subagent_type,
+            run_manifest,
+        )
     }
 
-    fn ensure_deep_review_retry_timeout(
-        input: &Value,
-        packet: &Value,
+    fn is_deep_review_auto_retry(input: &Value) -> bool {
+        input
+            .get("auto_retry")
+            .and_then(Value::as_bool)
+            .unwrap_or(false)
+    }
+
+    fn auto_retry_suppression_reason(code: &str) -> &'static str {
+        match code {
+            "deep_review_auto_retry_disabled" => "auto_retry_disabled",
+            "deep_review_auto_retry_elapsed_guard_exceeded" => "elapsed_guard_exceeded",
+            "deep_review_retry_budget_exhausted" => "budget_exhausted",
+            "deep_review_retry_without_initial_attempt" => "without_initial_attempt",
+            "deep_review_retry_missing_coverage" => "missing_coverage",
+            "deep_review_retry_missing_packet_id" => "missing_coverage",
+            "deep_review_retry_missing_status" => "missing_coverage",
+            "deep_review_retry_non_retryable_status" => "non_retryable_status",
+            "deep_review_retry_unknown_packet" => "unknown_packet",
+            "deep_review_retry_missing_packet_scope" => "unknown_packet",
+            "deep_review_retry_timeout_required" => "timeout_not_reduced",
+            "deep_review_retry_timeout_not_reduced" => "timeout_not_reduced",
+            "deep_review_retry_empty_scope" => "empty_scope",
+            "deep_review_retry_scope_not_reduced" => "scope_not_reduced",
+            _ => "invalid_coverage",
+        }
+    }
+
+    fn ensure_deep_review_auto_retry_allowed(
+        conc_policy: &DeepReviewConcurrencyPolicy,
+        dialog_turn_id: &str,
     ) -> Result<(), DeepReviewPolicyViolation> {
-        let retry_timeout_seconds =
-            Self::u64_for_any_key(input, &["timeout_seconds", "timeoutSeconds"]).unwrap_or(0);
-        if retry_timeout_seconds == 0 {
+        if !conc_policy.allow_bounded_auto_retry {
             return Err(DeepReviewPolicyViolation::new(
-                "deep_review_retry_timeout_required",
-                "DeepReview retry requires a positive timeout_seconds value",
+                "deep_review_auto_retry_disabled",
+                "DeepReview bounded automatic retry is disabled by Review Team settings",
             ));
         }
 
-        let source_timeout_seconds =
-            Self::u64_for_any_key(packet, &["timeoutSeconds", "timeout_seconds"]).unwrap_or(0);
-        if source_timeout_seconds > 0 && retry_timeout_seconds >= source_timeout_seconds {
-            return Err(DeepReviewPolicyViolation::new(
-                "deep_review_retry_timeout_not_reduced",
-                format!(
-                    "DeepReview retry timeout_seconds ({}) must be lower than source timeout ({})",
-                    retry_timeout_seconds, source_timeout_seconds
-                ),
-            ));
+        if let Some(elapsed_seconds) = deep_review_turn_elapsed_seconds(dialog_turn_id) {
+            if elapsed_seconds > conc_policy.auto_retry_elapsed_guard_seconds {
+                return Err(DeepReviewPolicyViolation::new(
+                    "deep_review_auto_retry_elapsed_guard_exceeded",
+                    format!(
+                        "DeepReview automatic retry elapsed guard exceeded (elapsed: {}s, guard: {}s)",
+                        elapsed_seconds, conc_policy.auto_retry_elapsed_guard_seconds
+                    ),
+                ));
+            }
         }
 
         Ok(())
     }
 
     fn prompt_with_deep_review_retry_scope(prompt: &str, retry_scope_files: &[String]) -> String {
-        let mut scoped_prompt = String::new();
-        scoped_prompt.push_str("<deep_review_retry_scope>\n");
-        scoped_prompt.push_str(
-            "This is a bounded DeepReview retry. Review only the following retry_scope_files and treat any other files as background context only:\n",
-        );
-        for file in retry_scope_files {
-            scoped_prompt.push_str("- ");
-            scoped_prompt.push_str(file);
-            scoped_prompt.push('\n');
-        }
-        scoped_prompt.push_str("</deep_review_retry_scope>\n\n");
-        scoped_prompt.push_str(prompt);
-        scoped_prompt
+        deep_review_task_adapter::prompt_with_deep_review_retry_scope(prompt, retry_scope_files)
     }
 
-    fn queue_reason_to_event_reason(
-        reason: DeepReviewCapacityQueueReason,
-    ) -> DeepReviewQueueReason {
-        match reason {
-            DeepReviewCapacityQueueReason::ProviderRateLimit => {
-                DeepReviewQueueReason::ProviderRateLimit
-            }
-            DeepReviewCapacityQueueReason::ProviderConcurrencyLimit => {
-                DeepReviewQueueReason::ProviderConcurrencyLimit
-            }
-            DeepReviewCapacityQueueReason::RetryAfter => DeepReviewQueueReason::RetryAfter,
-            DeepReviewCapacityQueueReason::LocalConcurrencyCap => {
-                DeepReviewQueueReason::LocalConcurrencyCap
-            }
-            DeepReviewCapacityQueueReason::TemporaryOverload => {
-                DeepReviewQueueReason::TemporaryOverload
-            }
-        }
-    }
-
-    fn queue_reason_to_snake_case(reason: DeepReviewCapacityQueueReason) -> &'static str {
-        match reason {
-            DeepReviewCapacityQueueReason::ProviderRateLimit => "provider_rate_limit",
-            DeepReviewCapacityQueueReason::ProviderConcurrencyLimit => "provider_concurrency_limit",
-            DeepReviewCapacityQueueReason::RetryAfter => "retry_after",
-            DeepReviewCapacityQueueReason::LocalConcurrencyCap => "local_concurrency_cap",
-            DeepReviewCapacityQueueReason::TemporaryOverload => "temporary_overload",
-        }
-    }
-
-    fn deep_review_capacity_reason_for_provider_error(
+    fn deep_review_capacity_decision_for_provider_error(
         error: &BitFunError,
-    ) -> Option<DeepReviewCapacityQueueReason> {
-        let detail = error.error_detail();
-        let error_message = error.to_string();
-        let code = detail.provider_code.as_deref().unwrap_or_default();
-        let message = detail
-            .provider_message
-            .as_deref()
-            .unwrap_or(error_message.as_str());
-        let decision = classify_deep_review_capacity_error(code, message, None);
-        if decision.queueable {
-            return decision.reason;
-        }
-
-        match detail.category {
-            ErrorCategory::RateLimit => Some(DeepReviewCapacityQueueReason::ProviderRateLimit),
-            ErrorCategory::ProviderUnavailable => {
-                Some(DeepReviewCapacityQueueReason::TemporaryOverload)
-            }
-            _ => None,
-        }
+    ) -> crate::agentic::deep_review_policy::DeepReviewCapacityQueueDecision {
+        deep_review_task_adapter::capacity_decision_for_provider_error(error)
     }
 
     fn deep_review_capacity_skip_result_for_provider_reason(
@@ -461,30 +162,77 @@ impl TaskTool {
         conc_policy: &DeepReviewConcurrencyPolicy,
         duration_ms: u128,
     ) -> (Value, String) {
-        let snapshot = record_deep_review_effective_concurrency_capacity_error(
-            dialog_turn_id,
-            conc_policy.max_parallel_instances,
+        deep_review_task_adapter::capacity_skip_result_for_provider_reason(
             reason,
-            None,
-        );
-        record_deep_review_capacity_skip(dialog_turn_id);
+            dialog_turn_id,
+            subagent_type,
+            conc_policy,
+            duration_ms,
+        )
+    }
 
-        let duration_ms = u64::try_from(duration_ms).unwrap_or(u64::MAX);
-        let reason_code = Self::queue_reason_to_snake_case(reason);
-        let assistant_message = format!(
-            "Subagent '{}' was skipped because the provider reported transient DeepReview capacity pressure.\n<queue_result status=\"capacity_skipped\" reason=\"{}\" queue_elapsed_ms=\"0\" />",
-            subagent_type, reason_code
-        );
-        let data = json!({
-            "duration": duration_ms,
-            "status": "capacity_skipped",
-            "queue_elapsed_ms": 0,
-            "max_queue_wait_seconds": conc_policy.max_queue_wait_seconds,
-            "queue_skip_reason": reason_code,
-            "effective_parallel_instances": snapshot.effective_parallel_instances
-        });
+    fn deep_review_capacity_skip_result_for_provider_queue_outcome(
+        reason: DeepReviewCapacityQueueReason,
+        dialog_turn_id: &str,
+        subagent_type: &str,
+        conc_policy: &DeepReviewConcurrencyPolicy,
+        duration_ms: u128,
+        queue_elapsed_ms: u64,
+        terminal_skip_reason: Option<DeepReviewQueueWaitSkipReason>,
+    ) -> (Value, String) {
+        deep_review_task_adapter::capacity_skip_result_for_provider_queue_outcome(
+            reason,
+            dialog_turn_id,
+            subagent_type,
+            conc_policy,
+            duration_ms,
+            queue_elapsed_ms,
+            terminal_skip_reason,
+        )
+    }
 
-        (data, assistant_message)
+    fn deep_review_provider_capacity_queue_wait_seconds(
+        decision: &crate::agentic::deep_review_policy::DeepReviewCapacityQueueDecision,
+        conc_policy: &DeepReviewConcurrencyPolicy,
+    ) -> Option<u64> {
+        deep_review_task_adapter::provider_capacity_queue_wait_seconds(decision, conc_policy)
+    }
+
+    async fn wait_for_deep_review_provider_capacity_retry(
+        session_id: &str,
+        dialog_turn_id: &str,
+        tool_id: &str,
+        subagent_type: &str,
+        conc_policy: &DeepReviewConcurrencyPolicy,
+        reason: DeepReviewCapacityQueueReason,
+        max_wait_seconds: u64,
+        is_optional_reviewer: bool,
+    ) -> DeepReviewProviderQueueWaitOutcome {
+        deep_review_task_adapter::wait_for_provider_capacity_retry(
+            session_id,
+            dialog_turn_id,
+            tool_id,
+            subagent_type,
+            conc_policy,
+            reason,
+            max_wait_seconds,
+            is_optional_reviewer,
+        )
+        .await
+    }
+
+    fn record_deep_review_provider_capacity_retry(
+        dialog_turn_id: &str,
+        reason: DeepReviewCapacityQueueReason,
+    ) {
+        deep_review_task_adapter::record_provider_capacity_retry(dialog_turn_id, reason);
+    }
+
+    fn record_deep_review_provider_capacity_retry_success(
+        dialog_turn_id: &str,
+        reason: DeepReviewCapacityQueueReason,
+    ) {
+        deep_review_task_adapter::record_provider_capacity_retry_success(dialog_turn_id, reason);
     }
 
     async fn emit_deep_review_queue_state(
@@ -501,29 +249,21 @@ impl TaskTool {
         queue_elapsed_ms: u64,
         max_queue_wait_seconds: u64,
     ) {
-        let run_elapsed_ms = matches!(&status, DeepReviewQueueStatus::Running).then_some(0);
-        if let Some(coordinator) = get_global_coordinator() {
-            coordinator
-                .emit_deep_review_queue_state_changed(
-                    session_id,
-                    dialog_turn_id,
-                    DeepReviewQueueState {
-                        tool_id: tool_id.to_string(),
-                        subagent_type: subagent_type.to_string(),
-                        status,
-                        reason: reason.map(Self::queue_reason_to_event_reason),
-                        queued_reviewer_count,
-                        active_reviewer_count: Some(active_reviewer_count),
-                        effective_parallel_instances,
-                        optional_reviewer_count,
-                        queue_elapsed_ms: Some(queue_elapsed_ms),
-                        run_elapsed_ms,
-                        max_queue_wait_seconds: Some(max_queue_wait_seconds),
-                        session_concurrency_high: false,
-                    },
-                )
-                .await;
-        }
+        deep_review_task_adapter::emit_queue_state(
+            session_id,
+            dialog_turn_id,
+            tool_id,
+            subagent_type,
+            status,
+            reason,
+            queued_reviewer_count,
+            active_reviewer_count,
+            optional_reviewer_count,
+            effective_parallel_instances,
+            queue_elapsed_ms,
+            max_queue_wait_seconds,
+        )
+        .await;
     }
 
     async fn wait_for_deep_review_reviewer_capacity(
@@ -534,168 +274,61 @@ impl TaskTool {
         conc_policy: &DeepReviewConcurrencyPolicy,
         is_optional_reviewer: bool,
     ) -> BitFunResult<DeepReviewQueueWaitOutcome> {
-        let decision = classify_deep_review_capacity_error(
-            "deep_review_concurrency_cap_reached",
-            "Maximum parallel reviewer instances reached",
-            None,
-        );
-        let reason = decision
-            .reason
-            .unwrap_or(DeepReviewCapacityQueueReason::LocalConcurrencyCap);
-        let started_at = Instant::now();
-        let max_wait = Duration::from_secs(conc_policy.max_queue_wait_seconds);
-        let mut paused_since: Option<Instant> = None;
-        let mut paused_total = Duration::ZERO;
-        let optional_reviewer_count = is_optional_reviewer.then_some(1);
+        deep_review_task_adapter::wait_for_reviewer_capacity(
+            session_id,
+            dialog_turn_id,
+            tool_id,
+            subagent_type,
+            conc_policy,
+            is_optional_reviewer,
+        )
+        .await
+    }
 
-        loop {
-            let now = Instant::now();
-            let current_pause_elapsed = paused_since
-                .map(|paused_at| now.saturating_duration_since(paused_at))
-                .unwrap_or_default();
-            let queue_elapsed = now
-                .saturating_duration_since(started_at)
-                .saturating_sub(paused_total)
-                .saturating_sub(current_pause_elapsed);
-            let queue_elapsed_ms = u64::try_from(queue_elapsed.as_millis()).unwrap_or(u64::MAX);
-            let active_reviewers = deep_review_active_reviewer_count(dialog_turn_id);
-            let effective_parallel_instances = deep_review_effective_parallel_instances(
-                dialog_turn_id,
-                conc_policy.max_parallel_instances,
-            );
+    fn deep_review_local_capacity_skip_tool_result(
+        dialog_turn_id: &str,
+        subagent_type: &str,
+        conc_policy: &DeepReviewConcurrencyPolicy,
+        skip_reason: DeepReviewQueueWaitSkipReason,
+        queue_elapsed_ms: u64,
+        duration_ms: u128,
+    ) -> ToolResult {
+        let queue_skip_reason = match skip_reason {
+            DeepReviewQueueWaitSkipReason::QueueExpired => "queue_expired",
+            DeepReviewQueueWaitSkipReason::UserCancelled => "user_cancelled",
+            DeepReviewQueueWaitSkipReason::OptionalSkipped => "optional_skipped",
+        };
+        let assistant_message = match skip_reason {
+            DeepReviewQueueWaitSkipReason::QueueExpired => format!(
+                "Subagent '{}' was skipped because the DeepReview capacity queue reached its maximum wait ({}s).\n<queue_result status=\"capacity_skipped\" reason=\"local_concurrency_cap\" queue_elapsed_ms=\"{}\" />",
+                subagent_type,
+                conc_policy.max_queue_wait_seconds,
+                queue_elapsed_ms
+            ),
+            DeepReviewQueueWaitSkipReason::UserCancelled => format!(
+                "Subagent '{}' was skipped because the DeepReview capacity queue was cancelled by the user.\n<queue_result status=\"capacity_skipped\" reason=\"user_cancelled\" queue_elapsed_ms=\"{}\" />",
+                subagent_type, queue_elapsed_ms
+            ),
+            DeepReviewQueueWaitSkipReason::OptionalSkipped => format!(
+                "Subagent '{}' was skipped because optional DeepReview queued reviewers were skipped by the user.\n<queue_result status=\"capacity_skipped\" reason=\"optional_skipped\" queue_elapsed_ms=\"{}\" />",
+                subagent_type, queue_elapsed_ms
+            ),
+        };
 
-            let control_snapshot = deep_review_queue_control_snapshot(dialog_turn_id, tool_id);
-            if control_snapshot.cancelled
-                || (is_optional_reviewer && control_snapshot.skip_optional)
-            {
-                record_deep_review_runtime_queue_wait(dialog_turn_id, queue_elapsed_ms);
-                record_deep_review_capacity_skip(dialog_turn_id);
-                clear_deep_review_queue_control_for_tool(dialog_turn_id, tool_id);
-                Self::emit_deep_review_queue_state(
-                    session_id,
-                    dialog_turn_id,
-                    tool_id,
-                    subagent_type,
-                    DeepReviewQueueStatus::CapacitySkipped,
-                    Some(reason),
-                    0,
-                    active_reviewers,
-                    optional_reviewer_count,
-                    Some(effective_parallel_instances),
-                    queue_elapsed_ms,
-                    conc_policy.max_queue_wait_seconds,
-                )
-                .await;
-                return Ok(DeepReviewQueueWaitOutcome::Skipped {
-                    queue_elapsed_ms,
-                    skip_reason: if control_snapshot.cancelled {
-                        DeepReviewQueueWaitSkipReason::UserCancelled
-                    } else {
-                        DeepReviewQueueWaitSkipReason::OptionalSkipped
-                    },
-                });
-            }
-
-            if control_snapshot.paused {
-                if paused_since.is_none() {
-                    paused_since = Some(now);
-                }
-                Self::emit_deep_review_queue_state(
-                    session_id,
-                    dialog_turn_id,
-                    tool_id,
-                    subagent_type,
-                    DeepReviewQueueStatus::PausedByUser,
-                    Some(reason),
-                    1,
-                    active_reviewers,
-                    optional_reviewer_count,
-                    Some(effective_parallel_instances),
-                    queue_elapsed_ms,
-                    conc_policy.max_queue_wait_seconds,
-                )
-                .await;
-                sleep(DEEP_REVIEW_QUEUE_POLL_INTERVAL).await;
-                continue;
-            }
-
-            if let Some(paused_at) = paused_since.take() {
-                paused_total += now.saturating_duration_since(paused_at);
-            }
-
-            if let Some(guard) =
-                try_begin_deep_review_active_reviewer(dialog_turn_id, effective_parallel_instances)
-            {
-                let active_reviewer_count = deep_review_active_reviewer_count(dialog_turn_id);
-                record_deep_review_runtime_queue_wait(dialog_turn_id, queue_elapsed_ms);
-                clear_deep_review_queue_control_for_tool(dialog_turn_id, tool_id);
-                Self::emit_deep_review_queue_state(
-                    session_id,
-                    dialog_turn_id,
-                    tool_id,
-                    subagent_type,
-                    DeepReviewQueueStatus::Running,
-                    None,
-                    0,
-                    active_reviewer_count,
-                    optional_reviewer_count,
-                    Some(effective_parallel_instances),
-                    queue_elapsed_ms,
-                    conc_policy.max_queue_wait_seconds,
-                )
-                .await;
-                return Ok(DeepReviewQueueWaitOutcome::Ready { guard });
-            }
-
-            if queue_elapsed >= max_wait {
-                let snapshot = record_deep_review_effective_concurrency_capacity_error(
+        ToolResult::Result {
+            data: json!({
+                "duration": duration_ms,
+                "status": "capacity_skipped",
+                "queue_elapsed_ms": queue_elapsed_ms,
+                "max_queue_wait_seconds": conc_policy.max_queue_wait_seconds,
+                "queue_skip_reason": queue_skip_reason,
+                "effective_parallel_instances": deep_review_effective_concurrency_snapshot(
                     dialog_turn_id,
                     conc_policy.max_parallel_instances,
-                    reason,
-                    decision.retry_after_seconds.map(Duration::from_secs),
-                );
-                record_deep_review_runtime_queue_wait(dialog_turn_id, queue_elapsed_ms);
-                record_deep_review_capacity_skip(dialog_turn_id);
-                clear_deep_review_queue_control_for_tool(dialog_turn_id, tool_id);
-                Self::emit_deep_review_queue_state(
-                    session_id,
-                    dialog_turn_id,
-                    tool_id,
-                    subagent_type,
-                    DeepReviewQueueStatus::CapacitySkipped,
-                    Some(reason),
-                    0,
-                    active_reviewers,
-                    optional_reviewer_count,
-                    Some(snapshot.effective_parallel_instances),
-                    queue_elapsed_ms,
-                    conc_policy.max_queue_wait_seconds,
-                )
-                .await;
-                return Ok(DeepReviewQueueWaitOutcome::Skipped {
-                    queue_elapsed_ms,
-                    skip_reason: DeepReviewQueueWaitSkipReason::QueueExpired,
-                });
-            }
-
-            Self::emit_deep_review_queue_state(
-                session_id,
-                dialog_turn_id,
-                tool_id,
-                subagent_type,
-                DeepReviewQueueStatus::QueuedForCapacity,
-                Some(reason),
-                1,
-                active_reviewers,
-                optional_reviewer_count,
-                Some(effective_parallel_instances),
-                queue_elapsed_ms,
-                conc_policy.max_queue_wait_seconds,
-            )
-            .await;
-
-            let remaining = max_wait.saturating_sub(queue_elapsed);
-            sleep(DEEP_REVIEW_QUEUE_POLL_INTERVAL.min(remaining)).await;
+                ).effective_parallel_instances
+            }),
+            result_for_assistant: Some(assistant_message),
+            image_attachments: None,
         }
     }
 
@@ -748,7 +381,7 @@ Usage notes:
 - The 'workspace_path' parameter must still be provided explicitly for the Explore and FileFinder agent.
 - Use 'model_id' when a caller needs a specific model or model slot for the subagent. Omit it to use the agent default.
 - Use 'timeout_seconds' when you need a hard deadline for the subagent. Omit it or set it to 0 to disable the timeout.
-- For DeepReview only, set 'retry' to true when re-dispatching a reviewer after that same reviewer returned partial_timeout or an explicit transient capacity failure in the current turn. Retry calls must include retry_coverage with source_packet_id, source_status, covered_files, and a smaller retry_scope_files list.
+- For DeepReview only, set 'retry' to true when re-dispatching a reviewer after that same reviewer returned partial_timeout or an explicit transient capacity failure in the current turn. Retry calls must include retry_coverage with source_packet_id, source_status, covered_files, and a smaller retry_scope_files list. Do not set 'auto_retry' unless this is a backend-owned automatic retry admitted by Review Team settings.
 - Launch multiple agents concurrently whenever possible, to maximize performance; to do that, use a single message with multiple tool calls
 - When the agent is done, it will return a single message back to you.
 - The agent's outputs should generally be trusted
@@ -874,6 +507,10 @@ impl Tool for TaskTool {
                 "retry": {
                     "type": "boolean",
                     "description": "DeepReview only: true when this Task call is a retry for the same reviewer role after partial_timeout or an explicit transient capacity failure in the current turn."
+                },
+                "auto_retry": {
+                    "type": "boolean",
+                    "description": "DeepReview only: true only for backend-owned bounded automatic retries. Requires Review Team auto retry opt-in and retry=true. User/model-issued retry actions must omit this field or set it to false."
                 },
                 "retry_coverage": {
                     "type": "object",
@@ -1056,6 +693,8 @@ impl Tool for TaskTool {
             None => None,
         };
         let is_retry = input.get("retry").and_then(Value::as_bool).unwrap_or(false);
+        let requested_auto_retry = Self::is_deep_review_auto_retry(input);
+        let is_auto_retry = is_retry && requested_auto_retry;
         let current_workspace_path = context
             .workspace_root()
             .map(|path| path.to_string_lossy().into_owned());
@@ -1075,7 +714,7 @@ impl Tool for TaskTool {
                 ));
             }
 
-            // For remote workspaces, skip local filesystem validation — the path
+            // For remote workspaces, skip local filesystem validation - the path
             // exists on the remote server, not locally.
             if !context.is_remote() {
                 let path = std::path::Path::new(&workspace_path);
@@ -1199,6 +838,11 @@ impl Tool for TaskTool {
                     ))
                 })?;
             deep_review_subagent_role = Some(role);
+            if requested_auto_retry && !is_retry {
+                return Err(BitFunError::tool(
+                    "auto_retry requires retry=true for DeepReview Task calls".to_string(),
+                ));
+            }
             if let Some(gate) = run_manifest
                 .as_ref()
                 .and_then(DeepReviewRunManifestGate::from_value)
@@ -1210,20 +854,42 @@ impl Tool for TaskTool {
                     ))
                 })?;
             }
+            let conc_policy = policy
+                .concurrency_policy_from_manifest(run_manifest.as_ref().unwrap_or(&Value::Null));
+            deep_review_concurrency_policy = Some(conc_policy.clone());
             if is_retry && role == DeepReviewSubagentRole::Reviewer {
-                deep_review_retry_scope_files = Some(
-                    Self::ensure_deep_review_retry_coverage(
-                        input,
-                        &subagent_type,
-                        run_manifest.as_ref(),
-                    )
-                    .map_err(|violation| {
-                        BitFunError::tool(format!(
+                deep_review_retry_scope_files = Some(match Self::ensure_deep_review_retry_coverage(
+                    input,
+                    &subagent_type,
+                    run_manifest.as_ref(),
+                ) {
+                    Ok(retry_scope_files) => retry_scope_files,
+                    Err(violation) => {
+                        if is_auto_retry {
+                            record_deep_review_runtime_auto_retry_suppressed(
+                                &dialog_turn_id,
+                                Self::auto_retry_suppression_reason(violation.code),
+                            );
+                        }
+                        return Err(BitFunError::tool(format!(
                             "DeepReview Task policy violation: {}",
                             violation.to_tool_error_message()
-                        ))
-                    })?,
-                );
+                        )));
+                    }
+                });
+                if is_auto_retry {
+                    Self::ensure_deep_review_auto_retry_allowed(&conc_policy, &dialog_turn_id)
+                        .map_err(|violation| {
+                            record_deep_review_runtime_auto_retry_suppressed(
+                                &dialog_turn_id,
+                                Self::auto_retry_suppression_reason(violation.code),
+                            );
+                            BitFunError::tool(format!(
+                                "DeepReview Task policy violation: {}",
+                                violation.to_tool_error_message()
+                            ))
+                        })?;
+                }
             }
             let is_readonly = get_agent_registry()
                 .get_subagent_is_readonly(&subagent_type)
@@ -1286,9 +952,6 @@ impl Tool for TaskTool {
             }
 
             // Enforce dynamic concurrency policy from the run manifest.
-            let conc_policy = policy
-                .concurrency_policy_from_manifest(run_manifest.as_ref().unwrap_or(&Value::Null));
-            deep_review_concurrency_policy = Some(conc_policy.clone());
             match role {
                 DeepReviewSubagentRole::Reviewer => {
                     deep_review_reviewer_configured_max_parallel_instances =
@@ -1325,46 +988,16 @@ impl Tool for TaskTool {
                                 queue_elapsed_ms,
                                 skip_reason,
                             } => {
-                                let queue_skip_reason = match skip_reason {
-                                    DeepReviewQueueWaitSkipReason::QueueExpired => "queue_expired",
-                                    DeepReviewQueueWaitSkipReason::UserCancelled => {
-                                        "user_cancelled"
-                                    }
-                                    DeepReviewQueueWaitSkipReason::OptionalSkipped => {
-                                        "optional_skipped"
-                                    }
-                                };
-                                let assistant_message = match skip_reason {
-                                    DeepReviewQueueWaitSkipReason::QueueExpired => format!(
-                                        "Subagent '{}' was skipped because the DeepReview capacity queue reached its maximum wait ({}s).\n<queue_result status=\"capacity_skipped\" reason=\"local_concurrency_cap\" queue_elapsed_ms=\"{}\" />",
-                                        subagent_type,
-                                        conc_policy.max_queue_wait_seconds,
-                                        queue_elapsed_ms
+                                return Ok(vec![
+                                    Self::deep_review_local_capacity_skip_tool_result(
+                                        &dialog_turn_id,
+                                        &subagent_type,
+                                        &conc_policy,
+                                        skip_reason,
+                                        queue_elapsed_ms,
+                                        start_time.elapsed().as_millis(),
                                     ),
-                                    DeepReviewQueueWaitSkipReason::UserCancelled => format!(
-                                        "Subagent '{}' was skipped because the DeepReview capacity queue was cancelled by the user.\n<queue_result status=\"capacity_skipped\" reason=\"user_cancelled\" queue_elapsed_ms=\"{}\" />",
-                                        subagent_type, queue_elapsed_ms
-                                    ),
-                                    DeepReviewQueueWaitSkipReason::OptionalSkipped => format!(
-                                        "Subagent '{}' was skipped because optional DeepReview queued reviewers were skipped by the user.\n<queue_result status=\"capacity_skipped\" reason=\"optional_skipped\" queue_elapsed_ms=\"{}\" />",
-                                        subagent_type, queue_elapsed_ms
-                                    ),
-                                };
-                                return Ok(vec![ToolResult::Result {
-                                    data: json!({
-                                        "duration": start_time.elapsed().as_millis(),
-                                        "status": "capacity_skipped",
-                                        "queue_elapsed_ms": queue_elapsed_ms,
-                                        "max_queue_wait_seconds": conc_policy.max_queue_wait_seconds,
-                                        "queue_skip_reason": queue_skip_reason,
-                                        "effective_parallel_instances": deep_review_effective_concurrency_snapshot(
-                                            &dialog_turn_id,
-                                            conc_policy.max_parallel_instances,
-                                        ).effective_parallel_instances
-                                    }),
-                                    result_for_assistant: Some(assistant_message),
-                                    image_attachments: None,
-                                }]);
+                                ]);
                             }
                         }
                     }
@@ -1390,22 +1023,30 @@ impl Tool for TaskTool {
                 is_retry,
             )
             .map_err(|violation| {
+                if is_auto_retry {
+                    record_deep_review_runtime_auto_retry_suppressed(
+                        &dialog_turn_id,
+                        Self::auto_retry_suppression_reason(violation.code),
+                    );
+                }
                 BitFunError::tool(format!(
                     "DeepReview Task policy violation: {}",
                     violation.to_tool_error_message()
                 ))
             })?;
+            if is_retry && role == DeepReviewSubagentRole::Reviewer {
+                if is_auto_retry {
+                    record_deep_review_runtime_auto_retry(&dialog_turn_id);
+                } else {
+                    record_deep_review_runtime_manual_retry(&dialog_turn_id);
+                }
+            }
         }
 
         if let Some(retry_scope_files) = deep_review_retry_scope_files.as_ref() {
             prompt = Self::prompt_with_deep_review_retry_scope(&prompt, retry_scope_files);
         }
 
-        let parent_info = SubagentParentInfo {
-            tool_call_id: tool_call_id.clone(),
-            session_id: session_id.clone(),
-            dialog_turn_id: dialog_turn_id.clone(),
-        };
         let subagent_context = deep_review_subagent_role.map(|role| {
             let mut values = HashMap::new();
             values.insert(
@@ -1422,65 +1063,216 @@ impl Tool for TaskTool {
             );
             values
         });
-        let result = match coordinator
-            .execute_subagent(
-                subagent_type.clone(),
-                prompt,
-                parent_info,
-                Some(effective_workspace_path.clone()),
-                subagent_context,
-                context.cancellation_token.as_ref(),
-                model_id,
-                timeout_seconds,
-            )
-            .await
-        {
-            Ok(result) => result,
-            Err(error) => {
-                if matches!(
-                    deep_review_subagent_role,
-                    Some(DeepReviewSubagentRole::Reviewer)
-                ) {
-                    if let (Some(reason), Some(conc_policy)) = (
-                        Self::deep_review_capacity_reason_for_provider_error(&error),
-                        deep_review_concurrency_policy.as_ref(),
-                    ) {
-                        drop(deep_review_active_guard.take());
-                        let (data, assistant_message) =
-                            Self::deep_review_capacity_skip_result_for_provider_reason(
-                                reason,
-                                &dialog_turn_id,
-                                &subagent_type,
-                                conc_policy,
-                                start_time.elapsed().as_millis(),
-                            );
-                        let effective_parallel_instances = data
-                            .get("effective_parallel_instances")
-                            .and_then(Value::as_u64)
-                            .and_then(|value| usize::try_from(value).ok());
-                        Self::emit_deep_review_queue_state(
-                            &session_id,
+        let prepared_prompt = prompt;
+        let mut provider_capacity_retry_reason: Option<DeepReviewCapacityQueueReason> = None;
+        let mut provider_capacity_queue_elapsed_ms = 0_u64;
+        let result = loop {
+            let parent_info = SubagentParentInfo {
+                tool_call_id: tool_call_id.clone(),
+                session_id: session_id.clone(),
+                dialog_turn_id: dialog_turn_id.clone(),
+            };
+            let execution_result = coordinator
+                .execute_subagent(
+                    subagent_type.clone(),
+                    prepared_prompt.clone(),
+                    parent_info,
+                    Some(effective_workspace_path.clone()),
+                    subagent_context.clone(),
+                    context.cancellation_token.as_ref(),
+                    model_id.clone(),
+                    timeout_seconds,
+                )
+                .await;
+
+            match execution_result {
+                Ok(result) => {
+                    if let Some(reason) = provider_capacity_retry_reason {
+                        Self::record_deep_review_provider_capacity_retry_success(
                             &dialog_turn_id,
-                            &tool_call_id,
-                            &subagent_type,
-                            DeepReviewQueueStatus::CapacitySkipped,
-                            Some(reason),
-                            0,
-                            deep_review_active_reviewer_count(&dialog_turn_id),
-                            deep_review_is_optional_reviewer.then_some(1),
-                            effective_parallel_instances,
-                            0,
-                            conc_policy.max_queue_wait_seconds,
-                        )
-                        .await;
-                        return Ok(vec![ToolResult::Result {
-                            data,
-                            result_for_assistant: Some(assistant_message),
-                            image_attachments: None,
-                        }]);
+                            reason,
+                        );
                     }
+                    break result;
                 }
-                return Err(error);
+                Err(error) => {
+                    if matches!(
+                        deep_review_subagent_role,
+                        Some(DeepReviewSubagentRole::Reviewer)
+                    ) {
+                        if let Some(conc_policy) = deep_review_concurrency_policy.as_ref() {
+                            let decision =
+                                Self::deep_review_capacity_decision_for_provider_error(&error);
+                            if let Some(reason) =
+                                decision.queueable.then_some(decision.reason).flatten()
+                            {
+                                drop(deep_review_active_guard.take());
+
+                                if provider_capacity_retry_reason.is_some() {
+                                    let (data, assistant_message) = Self::deep_review_capacity_skip_result_for_provider_queue_outcome(
+                                        reason,
+                                        &dialog_turn_id,
+                                        &subagent_type,
+                                        conc_policy,
+                                        start_time.elapsed().as_millis(),
+                                        provider_capacity_queue_elapsed_ms,
+                                        None,
+                                    );
+                                    let effective_parallel_instances = data
+                                        .get("effective_parallel_instances")
+                                        .and_then(Value::as_u64)
+                                        .and_then(|value| usize::try_from(value).ok());
+                                    Self::emit_deep_review_queue_state(
+                                        &session_id,
+                                        &dialog_turn_id,
+                                        &tool_call_id,
+                                        &subagent_type,
+                                        DeepReviewQueueStatus::CapacitySkipped,
+                                        Some(reason),
+                                        0,
+                                        deep_review_active_reviewer_count(&dialog_turn_id),
+                                        deep_review_is_optional_reviewer.then_some(1),
+                                        effective_parallel_instances,
+                                        provider_capacity_queue_elapsed_ms,
+                                        conc_policy.max_queue_wait_seconds,
+                                    )
+                                    .await;
+                                    return Ok(vec![ToolResult::Result {
+                                        data,
+                                        result_for_assistant: Some(assistant_message),
+                                        image_attachments: None,
+                                    }]);
+                                }
+
+                                if let Some(max_wait_seconds) =
+                                    Self::deep_review_provider_capacity_queue_wait_seconds(
+                                        &decision,
+                                        conc_policy,
+                                    )
+                                {
+                                    match Self::wait_for_deep_review_provider_capacity_retry(
+                                        &session_id,
+                                        &dialog_turn_id,
+                                        &tool_call_id,
+                                        &subagent_type,
+                                        conc_policy,
+                                        reason,
+                                        max_wait_seconds,
+                                        deep_review_is_optional_reviewer,
+                                    )
+                                    .await
+                                    {
+                                        DeepReviewProviderQueueWaitOutcome::ReadyToRetry {
+                                            queue_elapsed_ms,
+                                        } => {
+                                            provider_capacity_queue_elapsed_ms = queue_elapsed_ms;
+                                            let effective_parallel_instances =
+                                                deep_review_effective_parallel_instances(
+                                                    &dialog_turn_id,
+                                                    conc_policy.max_parallel_instances,
+                                                );
+                                            if let Some(guard) =
+                                                try_begin_deep_review_active_reviewer(
+                                                    &dialog_turn_id,
+                                                    effective_parallel_instances,
+                                                )
+                                            {
+                                                deep_review_active_guard = Some(guard);
+                                            } else {
+                                                match Self::wait_for_deep_review_reviewer_capacity(
+                                                    &session_id,
+                                                    &dialog_turn_id,
+                                                    &tool_call_id,
+                                                    &subagent_type,
+                                                    conc_policy,
+                                                    deep_review_is_optional_reviewer,
+                                                )
+                                                .await?
+                                                {
+                                                    DeepReviewQueueWaitOutcome::Ready { guard } => {
+                                                        deep_review_active_guard = Some(guard);
+                                                    }
+                                                    DeepReviewQueueWaitOutcome::Skipped {
+                                                        queue_elapsed_ms,
+                                                        skip_reason,
+                                                    } => {
+                                                        return Ok(vec![Self::deep_review_local_capacity_skip_tool_result(
+                                                            &dialog_turn_id,
+                                                            &subagent_type,
+                                                            conc_policy,
+                                                            skip_reason,
+                                                            queue_elapsed_ms,
+                                                            start_time.elapsed().as_millis(),
+                                                        )]);
+                                                    }
+                                                }
+                                            }
+                                            provider_capacity_retry_reason = Some(reason);
+                                            Self::record_deep_review_provider_capacity_retry(
+                                                &dialog_turn_id,
+                                                reason,
+                                            );
+                                            continue;
+                                        }
+                                        DeepReviewProviderQueueWaitOutcome::Skipped {
+                                            queue_elapsed_ms,
+                                            skip_reason,
+                                        } => {
+                                            let (data, assistant_message) = Self::deep_review_capacity_skip_result_for_provider_queue_outcome(
+                                                reason,
+                                                &dialog_turn_id,
+                                                &subagent_type,
+                                                conc_policy,
+                                                start_time.elapsed().as_millis(),
+                                                queue_elapsed_ms,
+                                                Some(skip_reason),
+                                            );
+                                            return Ok(vec![ToolResult::Result {
+                                                data,
+                                                result_for_assistant: Some(assistant_message),
+                                                image_attachments: None,
+                                            }]);
+                                        }
+                                    }
+                                }
+
+                                let (data, assistant_message) =
+                                    Self::deep_review_capacity_skip_result_for_provider_reason(
+                                        reason,
+                                        &dialog_turn_id,
+                                        &subagent_type,
+                                        conc_policy,
+                                        start_time.elapsed().as_millis(),
+                                    );
+                                let effective_parallel_instances = data
+                                    .get("effective_parallel_instances")
+                                    .and_then(Value::as_u64)
+                                    .and_then(|value| usize::try_from(value).ok());
+                                Self::emit_deep_review_queue_state(
+                                    &session_id,
+                                    &dialog_turn_id,
+                                    &tool_call_id,
+                                    &subagent_type,
+                                    DeepReviewQueueStatus::CapacitySkipped,
+                                    Some(reason),
+                                    0,
+                                    deep_review_active_reviewer_count(&dialog_turn_id),
+                                    deep_review_is_optional_reviewer.then_some(1),
+                                    effective_parallel_instances,
+                                    0,
+                                    conc_policy.max_queue_wait_seconds,
+                                )
+                                .await;
+                                return Ok(vec![ToolResult::Result {
+                                    data,
+                                    result_for_assistant: Some(assistant_message),
+                                    image_attachments: None,
+                                }]);
+                            }
+                        }
+                    }
+                    return Err(error);
+                }
             }
         };
         if !result.is_partial_timeout() {
@@ -1503,7 +1295,11 @@ impl Tool for TaskTool {
         };
 
         // Build retry hint for deep review reviewer timeouts.
-        let retry_hint = if result.is_partial_timeout() && !is_retry {
+        let retry_hint = if Self::should_emit_deep_review_retry_guidance(
+            result.is_partial_timeout(),
+            is_retry,
+            deep_review_subagent_role,
+        ) {
             let retries_used = crate::agentic::deep_review_policy::deep_review_retries_used(
                 &dialog_turn_id,
                 &subagent_type,
@@ -1564,6 +1360,7 @@ mod tests {
         DeepReviewBudgetTracker, DeepReviewExecutionPolicy, DeepReviewSubagentRole,
     };
     use crate::agentic::tools::framework::Tool;
+    use crate::util::BitFunError;
     use serde_json::json;
 
     #[test]
@@ -1703,16 +1500,18 @@ mod tests {
             stagger_seconds: 0,
             max_queue_wait_seconds: 60,
             batch_extras_separately: true,
+            allow_bounded_auto_retry: false,
+            auto_retry_elapsed_guard_seconds: 180,
         };
-        // 0 active → allowed
+        // 0 active -> allowed
         assert!(policy
             .check_launch_allowed(0, DeepReviewSubagentRole::Reviewer, false)
             .is_ok());
-        // 1 active → allowed
+        // 1 active -> allowed
         assert!(policy
             .check_launch_allowed(1, DeepReviewSubagentRole::Reviewer, false)
             .is_ok());
-        // 2 active (at cap) → blocked
+        // 2 active (at cap) -> blocked
         assert!(policy
             .check_launch_allowed(2, DeepReviewSubagentRole::Reviewer, false)
             .is_err());
@@ -1727,6 +1526,8 @@ mod tests {
             stagger_seconds: 0,
             max_queue_wait_seconds: 60,
             batch_extras_separately: true,
+            allow_bounded_auto_retry: false,
+            auto_retry_elapsed_guard_seconds: 180,
         };
         let violation = policy
             .check_launch_allowed(2, DeepReviewSubagentRole::Reviewer, false)
@@ -1757,6 +1558,8 @@ mod tests {
             stagger_seconds: 0,
             max_queue_wait_seconds: 0,
             batch_extras_separately: true,
+            allow_bounded_auto_retry: false,
+            auto_retry_elapsed_guard_seconds: 180,
         };
 
         let outcome = TaskTool::wait_for_deep_review_reviewer_capacity(
@@ -1809,6 +1612,8 @@ mod tests {
             stagger_seconds: 0,
             max_queue_wait_seconds: 60,
             batch_extras_separately: true,
+            allow_bounded_auto_retry: false,
+            auto_retry_elapsed_guard_seconds: 180,
         };
 
         let outcome = TaskTool::wait_for_deep_review_reviewer_capacity(
@@ -1851,6 +1656,8 @@ mod tests {
             stagger_seconds: 0,
             max_queue_wait_seconds: 1,
             batch_extras_separately: true,
+            allow_bounded_auto_retry: false,
+            auto_retry_elapsed_guard_seconds: 180,
         };
         let turn_id_owned = turn_id.to_string();
         let tool_id_owned = tool_id.to_string();
@@ -1885,7 +1692,10 @@ mod tests {
         let diagnostics = deep_review_runtime_diagnostics_snapshot(turn_id)
             .expect("runtime diagnostics should record terminal queue wait");
         assert_eq!(diagnostics.queue_wait_count, 1);
-        assert!(diagnostics.queue_wait_total_ms >= 20);
+        assert_eq!(
+            diagnostics.queue_wait_total_ms,
+            diagnostics.queue_wait_max_ms
+        );
     }
 
     #[tokio::test]
@@ -1905,6 +1715,8 @@ mod tests {
             stagger_seconds: 0,
             max_queue_wait_seconds: 0,
             batch_extras_separately: true,
+            allow_bounded_auto_retry: false,
+            auto_retry_elapsed_guard_seconds: 180,
         };
         let turn_id_owned = turn_id.to_string();
         let tool_id_owned = tool_id.to_string();
@@ -1966,6 +1778,8 @@ mod tests {
             stagger_seconds: 0,
             max_queue_wait_seconds: 60,
             batch_extras_separately: true,
+            allow_bounded_auto_retry: false,
+            auto_retry_elapsed_guard_seconds: 180,
         };
 
         let outcome = TaskTool::wait_for_deep_review_reviewer_capacity(
@@ -1996,15 +1810,15 @@ mod tests {
         use crate::agentic::deep_review_policy::DeepReviewConcurrencyPolicy;
 
         let policy = DeepReviewConcurrencyPolicy::default();
-        // 1 active reviewer → judge blocked
+        // 1 active reviewer -> judge blocked
         assert!(policy
             .check_launch_allowed(1, DeepReviewSubagentRole::Judge, false)
             .is_err());
-        // 0 active reviewers, no judge pending → judge allowed
+        // 0 active reviewers, no judge pending -> judge allowed
         assert!(policy
             .check_launch_allowed(0, DeepReviewSubagentRole::Judge, false)
             .is_ok());
-        // 0 active reviewers, judge already pending → blocked
+        // 0 active reviewers, judge already pending -> blocked
         assert!(policy
             .check_launch_allowed(0, DeepReviewSubagentRole::Judge, true)
             .is_err());
@@ -2035,7 +1849,7 @@ mod tests {
                 "fingerprint": "fp-new"
             }
         });
-        // Fingerprint mismatch → cache should not match
+        // Fingerprint mismatch -> cache should not match
         assert!(!cache.matches_manifest(&manifest));
     }
 
@@ -2172,6 +1986,74 @@ mod tests {
             TaskTool::deep_review_retry_guidance_max_retries(Some(&policy), "nonexistent-turn"),
             2
         );
+    }
+
+    #[test]
+    fn deep_review_retry_guidance_only_applies_to_initial_reviewer_timeout() {
+        assert!(TaskTool::should_emit_deep_review_retry_guidance(
+            true,
+            false,
+            Some(DeepReviewSubagentRole::Reviewer)
+        ));
+        assert!(!TaskTool::should_emit_deep_review_retry_guidance(
+            true, false, None
+        ));
+        assert!(!TaskTool::should_emit_deep_review_retry_guidance(
+            true,
+            false,
+            Some(DeepReviewSubagentRole::Judge)
+        ));
+        assert!(!TaskTool::should_emit_deep_review_retry_guidance(
+            true,
+            true,
+            Some(DeepReviewSubagentRole::Reviewer)
+        ));
+        assert!(!TaskTool::should_emit_deep_review_retry_guidance(
+            false,
+            false,
+            Some(DeepReviewSubagentRole::Reviewer)
+        ));
+    }
+
+    #[test]
+    fn deep_review_auto_retry_requires_review_team_opt_in() {
+        use crate::agentic::deep_review_policy::DeepReviewConcurrencyPolicy;
+
+        let policy = DeepReviewConcurrencyPolicy {
+            max_parallel_instances: 4,
+            stagger_seconds: 0,
+            max_queue_wait_seconds: 60,
+            batch_extras_separately: true,
+            allow_bounded_auto_retry: false,
+            auto_retry_elapsed_guard_seconds: 180,
+        };
+
+        let violation =
+            TaskTool::ensure_deep_review_auto_retry_allowed(&policy, "turn-auto-retry-disabled")
+                .expect_err("auto retry must be disabled by default");
+
+        assert_eq!(violation.code, "deep_review_auto_retry_disabled");
+        assert_eq!(
+            TaskTool::auto_retry_suppression_reason(violation.code),
+            "auto_retry_disabled"
+        );
+    }
+
+    #[test]
+    fn deep_review_auto_retry_opt_in_allows_guarded_admission() {
+        use crate::agentic::deep_review_policy::DeepReviewConcurrencyPolicy;
+
+        let policy = DeepReviewConcurrencyPolicy {
+            max_parallel_instances: 4,
+            stagger_seconds: 0,
+            max_queue_wait_seconds: 60,
+            batch_extras_separately: true,
+            allow_bounded_auto_retry: true,
+            auto_retry_elapsed_guard_seconds: 180,
+        };
+
+        TaskTool::ensure_deep_review_auto_retry_allowed(&policy, "turn-auto-retry-enabled")
+            .expect("opted-in auto retry should pass the admission gate before budget checks");
     }
 
     #[test]
@@ -2335,12 +2217,18 @@ mod tests {
             stagger_seconds: 0,
             max_queue_wait_seconds: 30,
             batch_extras_separately: true,
+            allow_bounded_auto_retry: false,
+            auto_retry_elapsed_guard_seconds: 180,
         };
         let turn_id = "turn-provider-capacity-skip";
-        let reason = TaskTool::deep_review_capacity_reason_for_provider_error(&BitFunError::ai(
-            "Provider error: provider=openai, code=429, message=rate limit exceeded",
-        ))
-        .expect("provider rate limit should surface as capacity_skipped");
+        let decision =
+            TaskTool::deep_review_capacity_decision_for_provider_error(&BitFunError::ai(
+                "Provider error: provider=openai, code=429, message=rate limit exceeded",
+            ));
+        assert!(decision.queueable);
+        let reason = decision
+            .reason
+            .expect("provider rate limit should surface as capacity_skipped");
         let (data, assistant_message) =
             TaskTool::deep_review_capacity_skip_result_for_provider_reason(
                 reason,
@@ -2365,14 +2253,175 @@ mod tests {
     fn deep_review_provider_quota_error_is_not_capacity_skipped() {
         use crate::util::BitFunError;
 
-        let reason = TaskTool::deep_review_capacity_reason_for_provider_error(&BitFunError::ai(
-            "Provider error: provider=glm, code=1113, message=insufficient quota",
-        ));
+        let decision = TaskTool::deep_review_capacity_decision_for_provider_error(
+            &BitFunError::ai("Provider error: provider=glm, code=1113, message=insufficient quota"),
+        );
 
         assert!(
-            reason.is_none(),
+            !decision.queueable,
             "quota errors should remain fail-fast instead of entering capacity queue flow"
         );
+    }
+
+    #[test]
+    fn deep_review_provider_queue_wait_is_bounded_by_retry_after_and_policy() {
+        use crate::agentic::deep_review_policy::DeepReviewConcurrencyPolicy;
+
+        let policy = DeepReviewConcurrencyPolicy {
+            max_parallel_instances: 3,
+            stagger_seconds: 0,
+            max_queue_wait_seconds: 30,
+            batch_extras_separately: true,
+            allow_bounded_auto_retry: false,
+            auto_retry_elapsed_guard_seconds: 180,
+        };
+        let decision = TaskTool::deep_review_capacity_decision_for_provider_error(
+            &BitFunError::ai("Provider error: code=429, message=Retry-After: 45"),
+        );
+
+        assert_eq!(
+            TaskTool::deep_review_provider_capacity_queue_wait_seconds(&decision, &policy),
+            Some(30)
+        );
+    }
+
+    #[test]
+    fn deep_review_provider_queue_wait_rejects_fail_fast_errors() {
+        use crate::agentic::deep_review_policy::DeepReviewConcurrencyPolicy;
+
+        let policy = DeepReviewConcurrencyPolicy {
+            max_parallel_instances: 3,
+            stagger_seconds: 0,
+            max_queue_wait_seconds: 30,
+            batch_extras_separately: true,
+            allow_bounded_auto_retry: false,
+            auto_retry_elapsed_guard_seconds: 180,
+        };
+        let decision = TaskTool::deep_review_capacity_decision_for_provider_error(
+            &BitFunError::ai("Provider error: code=invalid_model, message=model does not exist"),
+        );
+
+        assert_eq!(
+            TaskTool::deep_review_provider_capacity_queue_wait_seconds(&decision, &policy),
+            None
+        );
+    }
+
+    #[tokio::test]
+    async fn deep_review_provider_capacity_queue_cancel_control_skips_retry() {
+        use crate::agentic::deep_review::task_adapter::{
+            DeepReviewProviderQueueWaitOutcome, DeepReviewQueueWaitSkipReason,
+        };
+        use crate::agentic::deep_review_policy::{
+            apply_deep_review_queue_control, deep_review_runtime_diagnostics_snapshot,
+            DeepReviewCapacityQueueReason, DeepReviewConcurrencyPolicy,
+            DeepReviewQueueControlAction,
+        };
+
+        let turn_id = "turn-provider-queue-cancel";
+        let tool_id = "tool-provider-queue-cancel";
+        apply_deep_review_queue_control(turn_id, tool_id, DeepReviewQueueControlAction::Cancel);
+        let policy = DeepReviewConcurrencyPolicy {
+            max_parallel_instances: 2,
+            stagger_seconds: 0,
+            max_queue_wait_seconds: 60,
+            batch_extras_separately: true,
+            allow_bounded_auto_retry: false,
+            auto_retry_elapsed_guard_seconds: 180,
+        };
+
+        let outcome = TaskTool::wait_for_deep_review_provider_capacity_retry(
+            "session-provider-queue-cancel",
+            turn_id,
+            tool_id,
+            "ReviewSecurity",
+            &policy,
+            DeepReviewCapacityQueueReason::ProviderRateLimit,
+            60,
+            false,
+        )
+        .await;
+
+        match outcome {
+            DeepReviewProviderQueueWaitOutcome::Skipped {
+                queue_elapsed_ms,
+                skip_reason,
+            } => {
+                assert!(queue_elapsed_ms < 100);
+                assert_eq!(skip_reason, DeepReviewQueueWaitSkipReason::UserCancelled);
+            }
+            DeepReviewProviderQueueWaitOutcome::ReadyToRetry { .. } => {
+                panic!("cancelled provider queue should not retry")
+            }
+        }
+
+        let diagnostics = deep_review_runtime_diagnostics_snapshot(turn_id)
+            .expect("provider queue should record diagnostics");
+        assert_eq!(diagnostics.provider_capacity_queue_count, 1);
+        assert_eq!(
+            diagnostics
+                .provider_capacity_queue_reason_counts
+                .get("provider_rate_limit"),
+            Some(&1)
+        );
+    }
+
+    #[tokio::test]
+    async fn deep_review_provider_capacity_queue_pause_does_not_count_against_wait() {
+        use crate::agentic::deep_review::task_adapter::DeepReviewProviderQueueWaitOutcome;
+        use crate::agentic::deep_review_policy::{
+            apply_deep_review_queue_control, DeepReviewCapacityQueueReason,
+            DeepReviewConcurrencyPolicy, DeepReviewQueueControlAction,
+        };
+
+        let turn_id = "turn-provider-queue-pause";
+        let tool_id = "tool-provider-queue-pause";
+        apply_deep_review_queue_control(turn_id, tool_id, DeepReviewQueueControlAction::Pause);
+        let policy = DeepReviewConcurrencyPolicy {
+            max_parallel_instances: 2,
+            stagger_seconds: 0,
+            max_queue_wait_seconds: 1,
+            batch_extras_separately: true,
+            allow_bounded_auto_retry: false,
+            auto_retry_elapsed_guard_seconds: 180,
+        };
+        let turn_id_owned = turn_id.to_string();
+        let tool_id_owned = tool_id.to_string();
+
+        let handle = tokio::spawn(async move {
+            TaskTool::wait_for_deep_review_provider_capacity_retry(
+                "session-provider-queue-pause",
+                &turn_id_owned,
+                &tool_id_owned,
+                "ReviewSecurity",
+                &policy,
+                DeepReviewCapacityQueueReason::ProviderConcurrencyLimit,
+                1,
+                false,
+            )
+            .await
+        });
+
+        tokio::time::sleep(tokio::time::Duration::from_millis(30)).await;
+        assert!(
+            !handle.is_finished(),
+            "paused provider queue should not expire before continue"
+        );
+
+        apply_deep_review_queue_control(turn_id, tool_id, DeepReviewQueueControlAction::Continue);
+        let outcome = tokio::time::timeout(tokio::time::Duration::from_millis(1500), handle)
+            .await
+            .expect("continued provider queue should finish")
+            .expect("spawned wait should not panic");
+
+        match outcome {
+            DeepReviewProviderQueueWaitOutcome::ReadyToRetry { queue_elapsed_ms } => {
+                assert!(queue_elapsed_ms >= 900);
+            }
+            DeepReviewProviderQueueWaitOutcome::Skipped { .. } => {
+                panic!("continued provider queue should retry after bounded wait")
+            }
+        }
     }
 
     #[test]
