+++ src/crates/core/src/agentic/tools/pipeline/tool_pipeline.rs
@@ -6,6 +6,7 @@
 use super::state_manager::ToolStateManager;
 use super::types::*;
 use crate::agentic::core::{ToolCall, ToolExecutionState, ToolResult as ModelToolResult};
+use crate::agentic::deep_review::tool_context;
 use crate::agentic::events::types::ToolEventData;
 use crate::agentic::tools::computer_use_host::ComputerUseHostRef;
 use crate::agentic::tools::framework::{ToolResult as FrameworkToolResult, ToolUseContext};
@@ -694,7 +695,11 @@ impl ToolPipeline {
         // Loop detection: refuse to execute the same tool call repeatedly with
         // identical arguments. Triggered on the (THRESHOLD + 1)-th consecutive
         // identical call within the per-session sliding window.
-        if self.check_and_record_tool_call(&task.context.session_id, &tool_name, &tool_args) {
+        if self.check_and_record_tool_call(
+            &task.context.session_id,
+            &tool_name,
+            &tool_args,
+        ) {
             let error_msg = format!(
                 "Tool-call loop blocked: '{}' was already called {} times in a row in this session with identical arguments. Refusing to execute this {}th identical call. Issue a different tool call, or stop tool-calling and respond to the user. If you wrote a file recently and want to continue it, its full content is already visible in your earlier tool_use message — use Edit with `old_string` taken from the end of that content; do not Read the file again.",
                 tool_name,
@@ -1265,51 +1270,20 @@ impl ToolPipeline {
                         );
                     }
                 }
-                if let Some(raw_manifest) =
-                    task.context.context_vars.get("deep_review_run_manifest")
-                {
-                    if let Ok(manifest) = serde_json::from_str::<serde_json::Value>(raw_manifest) {
-                        map.insert("deep_review_run_manifest".to_string(), manifest);
-                    }
-                }
-                if let Some(role) = task.context.context_vars.get("deep_review_subagent_role") {
-                    if !role.trim().is_empty() {
-                        map.insert(
-                            "deep_review_subagent_role".to_string(),
-                            serde_json::json!(role.trim()),
-                        );
-                    }
-                }
-                if let Some(subagent_type) =
-                    task.context.context_vars.get("deep_review_subagent_type")
-                {
-                    if !subagent_type.trim().is_empty() {
-                        map.insert(
-                            "deep_review_subagent_type".to_string(),
-                            serde_json::json!(subagent_type.trim()),
-                        );
-                    }
-                }
-                if map
-                    .get("deep_review_subagent_role")
-                    .and_then(serde_json::Value::as_str)
-                    .is_some_and(|role| role == "reviewer")
-                {
-                    if let Some(parent_info) = task.context.subagent_parent_info.as_ref() {
-                        map.insert(
-                            "deep_review_parent_tool_call_id".to_string(),
-                            serde_json::json!(parent_info.tool_call_id.clone()),
-                        );
-                        map.insert(
-                            "deep_review_parent_session_id".to_string(),
-                            serde_json::json!(parent_info.session_id.clone()),
-                        );
-                        map.insert(
-                            "deep_review_parent_dialog_turn_id".to_string(),
-                            serde_json::json!(parent_info.dialog_turn_id.clone()),
-                        );
-                    }
-                }
+                let deep_review_parent_context =
+                    task.context
+                        .subagent_parent_info
+                        .as_ref()
+                        .map(|parent_info| tool_context::DeepReviewToolParentContext {
+                            tool_call_id: parent_info.tool_call_id.as_str(),
+                            session_id: parent_info.session_id.as_str(),
+                            dialog_turn_id: parent_info.dialog_turn_id.as_str(),
+                        });
+                tool_context::append_tool_use_context_data(
+                    &task.context.context_vars,
+                    deep_review_parent_context,
+                    &mut map,
+                );
 
                 map
             },
